# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: ConvoMeow native API
  version: 0.1.0
  description: >-
    The v1 API covers WhatsApp linked-device accounts, pairing, conversations,
    text and media messages, contacts, profile photos, and stored media.
security:
  - bearerToken: []
  - browserSession: []
paths:
  /api/v1/notifications/config:
    get:
      operationId: getNotificationConfig
      summary: Read browser push availability
      responses:
        '200':
          description: Notification availability and public signing key.
          content:
            application/json:
              schema:
                type: object
                required: [enabled]
                properties:
                  enabled: {type: boolean}
                  public_key:
                    type: string
                    description: Base64url VAPID public key, present when notifications are enabled.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/subscriptions:
    post:
      operationId: registerNotificationSubscription
      summary: Register this browser for push notifications
      security:
        - browserSession: []
      description: >-
        Requires a browser session cookie and a same-origin request. The server derives
        ownership and expiry from the cookie. Matching endpoints and encryption keys
        reconcile the same browser after a new sign-in. Request bodies are limited to 4 KiB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserPushSubscription'
      responses:
        '200':
          description: Subscription saved; use its ID for activity frames and removal.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: {type: string}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: Browser notifications are disabled on this server.
  /api/v1/notifications/subscriptions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    delete:
      operationId: deleteNotificationSubscription
      summary: Remove a browser push subscription
      security:
        - browserSession: []
      responses:
        '204':
          description: Subscription and pending notifications removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/events:
    get:
      parameters:
        - name: presence_account_id
          in: query
          description: Keep this account online while the stream is open. The last viewer leaving marks it unavailable.
          schema: {type: string}
      operationId: streamEvents
      summary: Subscribe to live resource changes
      description: >-
        Upgrade to WebSocket with the existing browser session or bearer token.
        The stream sends a ready frame, then resource-change and transient presence notifications.
        Browser clients can send browser.activity frames to suppress push for a focused conversation.
        Refresh HTTP data after every connection; notifications are not replayed.
        See realtime.md for message types and connection lifecycle.
      responses:
        '101':
          description: WebSocket connection established.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: WebSocket origin is not allowed.
...
  /api/v1/accounts:
    get:
      operationId: listAccounts
      summary: List accounts
      responses:
        '200':
          description: Accounts in this installation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createAccount
      summary: Create a WhatsApp linked-device account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
            example:
              label: sales
              provider: whatsapp
              connection_kind: linked_device
      responses:
        '201':
          description: Account created.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{id}:
    parameters:
      - $ref: '#/components/parameters/AccountID'
    get:
      operationId: getAccount
      summary: Read an account
      responses:
        '200':
          description: Account and connection state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/accounts/{id}/login-attempts:
    parameters:
      - $ref: '#/components/parameters/AccountID'
    post:
      operationId: startLoginAttempt
      summary: Start WhatsApp pairing
      responses:
        '202':
          description: Pairing started. Poll the Location URL for a QR challenge.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginAttempt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{id}/uploads:
    parameters:
      - $ref: '#/components/parameters/AccountID'
    post:
      operationId: createUpload
      summary: Store one file for a media message
      description: Unused uploads expire after 24 hours and can be sent once.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File stored.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '503':
          description: Media storage or staging space is unavailable.
        '507':
          description: Media storage quota is full.
  /api/v1/accounts/{id}/uploads/{upload_id}:
    parameters:
      - $ref: '#/components/parameters/AccountID'
      - name: upload_id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteUpload
      summary: Delete an unused upload
      responses:
        '204':
          description: Upload marked for deletion.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/accounts/{id}/login-attempts/{attempt_id}:
    parameters:
      - $ref: '#/components/parameters/AccountID'
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getLoginAttempt
      summary: Read a pairing attempt
      responses:
        '200':
          description: Current pairing state and any active QR challenge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginAttempt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/accounts/{id}/conversations:
    parameters:
      - $ref: '#/components/parameters/AccountID'
    post:
      operationId: createConversation
      summary: Find or create a conversation for a phone number or synced contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversation'
            example:
              target:
                type: phone_number
                value: '+15551234567'
      responses:
        '200':
          description: Existing conversation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '201':
          description: Conversation created.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{id}/contacts:
    parameters:
      - $ref: '#/components/parameters/AccountID'
    get:
      operationId: listContacts
      summary: List synced contacts
      description: Includes contacts with no conversation. Results use contact name order.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          description: Match a name, phone number, or provider ID.
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: Contacts from the linked account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{id}/contacts/{provider_id}:
    parameters:
      - $ref: '#/components/parameters/AccountID'
      - $ref: '#/components/parameters/ProviderContactID'
    get:
      operationId: getContact
      summary: Read a synced contact
      responses:
        '200':
          description: Saved name and available phone number.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{id}/avatar:
    parameters:
      - $ref: '#/components/parameters/AccountID'
    get:
      operationId: getAccountAvatar
      summary: Read the account's own profile photo
      description: Returns a saved preview while the account is disconnected. A missing photo requires a connected account to fetch it.
      responses:
        '200':
          description: Profile photo preview.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          description: Profile photo could not be fetched from the provider.
        '503':
          description: Saved media is unavailable or staging is busy.
        '507':
          description: Media storage limit reached.
  /api/v1/accounts/{id}/contacts/{provider_id}/avatar:
    parameters:
      - $ref: '#/components/parameters/AccountID'
      - $ref: '#/components/parameters/ProviderContactID'
    get:
      operationId: getContactAvatar
      summary: Read a contact's profile photo
      description: Returns a saved preview while the account is disconnected. A missing photo requires a connected account to fetch it.
      responses:
        '200':
          description: Profile photo preview.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          description: Profile photo could not be fetched from the provider.
        '503':
          description: Saved media is unavailable or staging is busy.
        '507':
          description: Media storage limit reached.
  /api/v1/conversations:
    get:
      operationId: listConversations
      summary: List conversations
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
          description: Limit the list to one account.
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Newest conversations first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/conversations/{id}:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
    get:
      operationId: getConversation
      summary: Read a conversation
      responses:
        '200':
          description: Conversation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/conversations/{id}/avatar:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
    get:
      operationId: getConversationAvatar
      summary: Read a conversation's profile photo
      description: Returns a saved contact or group preview while the account is disconnected.
      responses:
        '200':
          description: Contact or group photo preview.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          description: Profile photo could not be fetched from the provider.
        '503':
          description: Saved media is unavailable or staging is busy.
        '507':
          description: Media storage limit reached.
  /api/v1/conversations/{id}/history-requests:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
    post:
      operationId: requestConversationHistory
      summary: Request earlier messages from the provider
      description: Requests up to 50 messages before a saved, confirmed message. Provider responses merge into the conversation asynchronously. Acceptance does not guarantee delivery or complete history; the primary device may need to be online.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [before_message_id]
              properties:
                before_message_id:
                  type: string
                count:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 50
      responses:
        '202':
          description: History request sent to the provider.
          content:
            application/json:
              schema:
                type: object
                required: [request_id, status]
                properties:
                  request_id:
                    type: string
                  status:
                    type: string
                    enum: [requested]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          description: History request failed.
...
  /api/v1/messages:
    get:
      operationId: listAccountMessages
      summary: List messages for one account
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Newest provider message timestamps first. History may arrive after newer messages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/conversations/{id}/presence:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    post:
      operationId: sendChatPresence
      summary: Send typing, recording, or paused activity
      description: >-
        Sends transient activity to the provider without saving a message.
        Throttle repeated activity updates and send paused on inactivity or leaving the chat.
        Active client leases expire after ten seconds; one client pausing does not cancel another.
        The server limits unchanged activity refreshes to once per three seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [activity, client_id]
              properties:
                client_id:
                  type: string
                  minLength: 1
                  maxLength: 128
                  description: Unique identifier for this composer instance; reuse it for refresh and paused commands.
                activity:
                  type: string
                  enum: [typing, recording, paused]
      responses:
        '204':
          description: Activity sent to the provider.
        '400':
          description: Invalid activity or unsupported provider capability.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Account is not connected.
...
  /api/v1/messages/{id}/reactions:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: listMessageReactions
      summary: List current reaction participants
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Current participants ordered by opaque participant ID. Removed reactions are omitted.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MessageReaction'
                  next_cursor: {type: string}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: convomeow_session
      description: Signed browser session from /app/session. Write requests require the same origin.
    bearerToken:
      type: http
      scheme: bearer
      description: >-
        The current control token grants access to every account. Keep this API
        on loopback until user sessions and scoped API keys are implemented.
...
  responses:
...
    Unauthorized:
      description: Required bearer token or browser session missing or invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    BrowserPushSubscription:
      type: object
      additionalProperties: false
      required: [endpoint, keys, locale, preview]
      properties:
        endpoint:
          type: string
          format: uri
          maxLength: 2048
          description: Public HTTPS push endpoint supplied by the browser; only port 443 is allowed.
        keys:
          type: object
          additionalProperties: false
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
              description: Base64url uncompressed P-256 public key.
            auth:
              type: string
              description: Base64url 16-byte authentication secret.
        locale:
          type: string
          enum: [en, pt-BR]
        preview:
          type: boolean
          description: Include message content and media type; account and conversation names remain visible.
...